PR feedback - reworking local session setup
