+++ sis-local/streamlit_in_snowflake/__init__.py
@@ -1,2 +1,2 @@
 from streamlit_in_snowflake.sis import *
-from streamlit_in_snowflake.local_connection import LocalSnowparkConnection
+from streamlit_in_snowflake import local_session
